+++ src/kernel/memory.c
@@ -1,6 +1,6 @@
 #include "kernel/memory.h"
 
-void *malloc(uint32_t size)
+void *malloc(size_t size)
 {
     return (void *)0xDEADBEEF; // Placeholder for malloc
 }
@@ -9,20 +9,20 @@ void free(void *ptr)
 {
 }
 
-void memset(void *ptr, uint32_t value, uint32_t num)
+void memset(void *ptr, uint8_t value, size_t num)
 {
     uint8_t *p = (uint8_t *)ptr;
-    for (uint32_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
     {
-        p[i] = (uint8_t)value;
+        p[i] = value;
     }
 }
 
-void *memcpy(const void *restrict _dest, const void *restrict _src, uint64_t num_bytes)
+void *memcpy(const void *restrict _dest, const void *restrict _src, size_t num_bytes)
 {
     uint8_t *src = _src;
     uint8_t *dest = _dest;
-    for (uint64_t i = 0; i < num_bytes; i++)
+    for (size_t i = 0; i < num_bytes; i++)
     {
         dest[i] = src[i];
     }
